Issue #1392: Replace hook processingi by using a class ImportHook.
